Add TTA to DeepEdit App
