+++ sample-apps/deepedit/validate_TTA.py
@@ -47,7 +47,7 @@ for idx in range(len(new_labels)):
 
     # Waiting to train model for some epochs
     for i in range(args.training_time_factor):
-        sleep(30)
+        sleep(3)
         print("Training - Label number: " + str(idx) + " of " + str(len(new_labels) + 1) + " -- Second: " + str(i + 1))
 
     # Stop training
