--- sample-apps/segmentation_spleen_tta/README.md
@@ -1,14 +1,0 @@
-# Segmentation - Spleen
-
-## Overview
-
-This is a reference app to run infer + train tasks to segment liver and tumor. It uses pre-trained weights/model (UNET)
-from NVIDIA Clara. This also show-cases an example for multi-label segmentation.
-
-> It also enables **NVIDIA Clara Deepgrow models** and actions are limited to infer only.
-
-## Links
-
-- https://ngc.nvidia.com/catalog/models/nvidia:med:clara_pt_spleen_ct_segmentation
-- https://ngc.nvidia.com/catalog/models/nvidia:med:clara_pt_deepgrow_3d_annotation
-- https://ngc.nvidia.com/catalog/models/nvidia:med:clara_pt_deepgrow_2d_annotation
